refactor: Remove i18n duplicate file
